# 18 optimize state searching algorithm