objectives:
- reduce execution time
- reduce stack usage
- reduce program space usage