@AntoineMugnier 2023-12-25
![SM_of_PerfTestSM](https://github.com/AntoineMugnier/kaori-hsm/assets/55192906/39dad286-44b8-468a-8772-2d8bab93ead6)

[Perf test used for benchmarking](https://github.com/AntoineMugnier/kaori-hsm-perf-test). A series of event have been sent to the HSM, and their processing time measured.

NEW ALGO (Release profile):
(HOST) INFO  program has used at least 0.53/17.92 KiB (3.0%) of stack space

arm-none-eabi-size:
   text	   data	    bss	    dec	    hex	filename
  15248	     60	   2068	  17376	   43e0	target/thumbv7m-none-eabi/release/kaori-hsm-perf-test
 
Init state machine
Kaori HSM init took 74 µs

Dispatching evt A
Kaori HSM evt dispatch took 58 µs

Dispatching evt B
Kaori HSM evt dispatch took 126 µs

Dispatching evt C
Kaori HSM evt dispatch took 65 µs

Dispatching evt D
Kaori HSM evt dispatch took 159 µs

Dispatching evt B
Kaori HSM evt dispatch took 127 µs

Dispatching evt C
Kaori HSM evt dispatch took 140 µs

Dispatching evt B
Kaori HSM evt dispatch took 144 µs


OLD ALGO (release profile):
(HOST) INFO  program has used at least 0.57/17.92 KiB (3.2%) of stack space

arm-none-eabi-size:
   text	   data	    bss	    dec	    hex	filename
  15184	     60	   2068	  17312	   43a0	target/thumbv7m-none-eabi/release/kaori-hsm-perf-test

Init state machine
Kaori HSM init took 74 µs

Dispatching evt A
Kaori HSM evt dispatch took 59 µs

Dispatching evt B
Kaori HSM evt dispatch took 164 µs

Dispatching evt C
Kaori HSM evt dispatch took 158 µs

Dispatching evt D
Kaori HSM evt dispatch took 208 µs

Dispatching evt B
Kaori HSM evt dispatch took 164 µs

Dispatching evt C
Kaori HSM evt dispatch took 175 µs

Dispatching evt B
Kaori HSM evt dispatch took 180 µs

QPCPP ALGO:

arm-none-eabi-size:
   text	   data	    bss	    dec	    hex	filename
  13432	     60	   2068	  15560	   3cc8	target/thumbv7m-none-eabi/release/kaori-hsm-perf-test

Init state machine
QPCPP HSM init took 84 µs

Dispatching evt A
QPCPP HSM evt dispatch took 54 µs

Dispatching evt B
QPCPP HSM evt dispatch took 84 µs

Dispatching evt C
QPCPP HSM evt dispatch took 59 µs

Dispatching evt D
QPCPP HSM evt dispatch took 110 µs

Dispatching evt B
QPCPP HSM evt dispatch took 84 µs

Dispatching evt C
QPCPP HSM evt dispatch took 111 µs

Dispatching evt B
QPCPP HSM evt dispatch took 83 µs


(HOST) INFO  program has used at least 0.34/17.92 KiB (1.9%) of stack space

Remarks:
- execution time improved significantly
- slight improvement of stack usage